Simpler returns

+++ Marlin/src/sd/SdBaseFile.cpp
@@ -47,9 +47,8 @@ void (*SdBaseFile::dateTime_)(uint16_t* date, uint16_t* time) = 0;
 
 // add a cluster to a file
 bool SdBaseFile::addCluster() {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   if (!vol_->allocContiguous(1, &curCluster_)) return false;
 
   // if first cluster of file link to directory entry
@@ -58,15 +57,13 @@ bool SdBaseFile::addCluster() {
     flags_ |= F_FILE_DIR_DIRTY;
   }
   return true;
-  #endif
 }
 
 // Add a cluster to a directory file and zero the cluster.
 // return with first block of cluster in the cache
 bool SdBaseFile::addDirCluster() {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint32_t block;
   // max folder size
   if (fileSize_ / sizeof(dir_t) >= 0xFFFF) return false;
@@ -89,7 +86,6 @@ bool SdBaseFile::addDirCluster() {
   // Increase directory file size by cluster size
   fileSize_ += 512UL << vol_->clusterSizeShift_;
   return true;
-  #endif
 }
 
 // cache a file's directory entry
@@ -161,9 +157,8 @@ bool SdBaseFile::contiguousRange(uint32_t* bgnBlock, uint32_t* endBlock) {
  *
  */
 bool SdBaseFile::createContiguous(SdBaseFile* dirFile, const char* path, uint32_t size) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint32_t count;
   // don't allow zero length file
   if (size == 0) return false;
@@ -183,7 +178,6 @@ bool SdBaseFile::createContiguous(SdBaseFile* dirFile, const char* path, uint32_
   flags_ |= F_FILE_DIR_DIRTY;
 
   return sync();
-  #endif
 }
 
 /**
@@ -431,9 +425,8 @@ bool SdBaseFile::make83Name(const char* str, uint8_t* name, const char** ptr) {
  * directory, \a path is invalid or already exists in \a parent.
  */
 bool SdBaseFile::mkdir(SdBaseFile* parent, const char* path, bool pFlag) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint8_t dname[11];
   SdBaseFile dir1, dir2;
   SdBaseFile* sub = &dir1;
@@ -461,13 +454,11 @@ bool SdBaseFile::mkdir(SdBaseFile* parent, const char* path, bool pFlag) {
     sub = parent != &dir1 ? &dir1 : &dir2;
   }
   return mkdir(parent, dname);
-  #endif
 }
 
 bool SdBaseFile::mkdir(SdBaseFile* parent, const uint8_t dname[11]) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint32_t block;
   dir_t d;
   dir_t* p;
@@ -521,7 +512,6 @@ bool SdBaseFile::mkdir(SdBaseFile* parent, const uint8_t dname[11]) {
 
   // write first block
   return vol_->cacheFlush();
-  #endif
 }
 
 /**
@@ -737,7 +727,7 @@ bool SdBaseFile::open(SdBaseFile* dirFile, uint16_t index, uint8_t oflag) {
 // open a cached directory entry. Assumes vol_ is initialized
 bool SdBaseFile::openCachedEntry(uint8_t dirIndex, uint8_t oflag) {
   #if ENABLED(SDCARD_READONLY)
-  if (oflag & (O_WRITE | O_CREAT | O_TRUNC)) goto FAIL;
+    if (oflag & (O_WRITE | O_CREAT | O_TRUNC)) goto FAIL;
   #endif
 
   // location of entry in cache
@@ -1159,9 +1149,8 @@ dir_t* SdBaseFile::readDirCache() {
  * or an I/O error occurred.
  */
 bool SdBaseFile::remove() {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   dir_t* d;
   // free any clusters - will fail if read-only or directory
   if (!truncate(0)) return false;
@@ -1179,7 +1168,6 @@ bool SdBaseFile::remove() {
   // write entry to SD
   return vol_->cacheFlush();
   return true;
-  #endif
 }
 
 /**
@@ -1200,12 +1188,10 @@ bool SdBaseFile::remove() {
  * or an I/O error occurred.
  */
 bool SdBaseFile::remove(SdBaseFile* dirFile, const char* path) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   SdBaseFile file;
   return file.open(dirFile, path, O_WRITE) ? file.remove() : false;
-  #endif
 }
 
 /**
@@ -1219,9 +1205,8 @@ bool SdBaseFile::remove(SdBaseFile* dirFile, const char* path) {
  * file, newPath is invalid or already exists, or an I/O error occurs.
  */
 bool SdBaseFile::rename(SdBaseFile* dirFile, const char* newPath) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   dir_t entry;
   uint32_t dirCluster = 0;
   SdBaseFile file;
@@ -1296,7 +1281,6 @@ restore:
     vol_->cacheFlush();
   }
   return false;
-  #endif
 }
 
 /**
@@ -1315,9 +1299,8 @@ restore:
  * directory, is not empty, or an I/O error occurred.
  */
 bool SdBaseFile::rmdir() {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   // must be open subdirectory
   if (!isSubDir()) return false;
 
@@ -1338,7 +1321,6 @@ bool SdBaseFile::rmdir() {
   type_ = FAT_FILE_TYPE_NORMAL;
   flags_ |= O_WRITE;
   return remove();
-  #endif
 }
 
 /**
@@ -1357,9 +1339,8 @@ bool SdBaseFile::rmdir() {
  * \return true for success, false for failure.
  */
 bool SdBaseFile::rmRfStar() {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint32_t index;
   SdBaseFile f;
   rewind();
@@ -1399,7 +1380,6 @@ bool SdBaseFile::rmRfStar() {
     if (!rmdir()) return false;
   }
   return true;
-  #endif
 }
 
 /**
@@ -1467,9 +1447,8 @@ void SdBaseFile::setpos(filepos_t* pos) {
  * opened or an I/O error.
  */
 bool SdBaseFile::sync() {
-  #if DISABLED(SDCARD_READONLY)
   // only allow open files and directories
-  if (!isOpen()) goto FAIL;
+  if (ENABLED(SDCARD_READONLY) || !isOpen()) goto FAIL;
 
   if (flags_ & F_FILE_DIR_DIRTY) {
     dir_t* d = cacheDirEntry(SdVolume::CACHE_FOR_WRITE);
@@ -1494,7 +1473,6 @@ bool SdBaseFile::sync() {
   return vol_->cacheFlush();
 
   FAIL:
-  #endif
   writeError = true;
   return false;
 }
@@ -1570,9 +1548,8 @@ bool SdBaseFile::timestamp(SdBaseFile* file) {
  */
 bool SdBaseFile::timestamp(uint8_t flags, uint16_t year, uint8_t month,
                            uint8_t day, uint8_t hour, uint8_t minute, uint8_t second) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint16_t dirDate, dirTime;
   dir_t* d;
 
@@ -1610,7 +1587,6 @@ bool SdBaseFile::timestamp(uint8_t flags, uint16_t year, uint8_t month,
     d->lastWriteTime = dirTime;
   }
   return vol_->cacheFlush();
-  #endif
 }
 
 /**
@@ -1625,9 +1601,8 @@ bool SdBaseFile::timestamp(uint8_t flags, uint16_t year, uint8_t month,
  * \a length is greater than the current file size or an I/O error occurs.
  */
 bool SdBaseFile::truncate(uint32_t length) {
-  #if ENABLED(SDCARD_READONLY)
-  return false;
-  #else
+  if (ENABLED(SDCARD_READONLY)) return false;
+
   uint32_t newPos;
   // error if not a normal file or read-only
   if (!isFile() || !(flags_ & O_WRITE)) return false;
@@ -1670,7 +1645,6 @@ bool SdBaseFile::truncate(uint32_t length) {
 
   // set file to correct position
   return seekSet(newPos);
-  #endif
 }
 
 /**
@@ -1690,7 +1664,10 @@ bool SdBaseFile::truncate(uint32_t length) {
  *
  */
 int16_t SdBaseFile::write(const void* buf, uint16_t nbyte) {
-  #if DISABLED(SDCARD_READONLY)
+  #if ENABLED(SDCARD_READONLY)
+    writeError = true; return -1;
+  #endif
+
   // convert void* to uint8_t*  -  must be before goto statements
   const uint8_t* src = reinterpret_cast<const uint8_t*>(buf);
 
@@ -1781,7 +1758,6 @@ int16_t SdBaseFile::write(const void* buf, uint16_t nbyte) {
   return nbyte;
 
   FAIL:
-  #endif
   // return for write error
   writeError = true;
   return -1;
